Add version awareness to the monitoring API
This commit enable the monitoring API to
  * filter a workflow's instances list by their version
  * read a specific version of a workflow's definition

The workflow directed graph caching is also refactored
within this commit.

+++ workflow-bot-app/src/main/java/com/symphony/bdk/workflow/monitoring/repository/domain/WorkflowDomain.java
@@ -10,5 +10,5 @@ import lombok.RequiredArgsConstructor;
 public class WorkflowDomain {
   private final String id;
   private final String name;
-  private final int version;
+  private final Long version;
 }
